% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_TLMR}
\alias{run_TLMR}
\title{Run TLMR Estimation}
\usage{
run_TLMR(
  exposure_data,
  outcome_data,
  p_threshold,
  covariates_type = 0,
  summary_cov = NULL,
  n_divides = 3
)
}
\arguments{
\item{exposure_data}{The exposure data. The data should be in list format, with variable names specified as follows: instrumental variables as Z, exposure as X, outcome as Y, and covariates as C.}

\item{outcome_data}{The outcome data. The data should be in list format, with variable names specified as follows: instrumental variables as Z, exposure as X, outcome as Y, and covariates as C.}

\item{p_threshold}{The setting of the p-value threshold used to test the significance of effect modification.The default is 0.005.}

\item{covariates_type}{0 represents continuous type, 1 represents classified type. The default is 0.}

\item{summary_cov}{Covariates that need to be adjusted for calculating summary data, vector format(i.e. c(1,2)). The default is NULL.}

\item{n_divides}{When using IPW or AIPW estimation, this specifies the number of categories used to discretize continuous covariates. The default is 3.}
}
\value{
A list with results.
}
\description{
This is the main function that runs the full TLMR procedure.
}
